+++ Bookish/Views/Book/SearchBooks.cshtml
@@ -7,11 +7,21 @@
             if (property.Name != "Id")
             {
                 <div class="mb-3">
-                    <label for=@property.Name class="form-label">@property.Name</label>
-                    <input type="text" class="form-control" id=@property.Name name=@property.Name>
+                    <label for="@property.Name" class="form-label">@property.Name</label>
+                    <input type="text" class="form-control" id="@property.Name" name="searchParameters.@property.Name">
                 </div>
             }
         }
+        <div class="mb-3">
+            <label for="orderBy" class="form-label">Order by</label>
+            <select class="form-select" id="orderBy" name="orderBy">
+                @foreach (var property in typeof(BookModel).GetProperties())
+                {
+                    <option value="@property.Name">@property.Name</option>
+                }
+            </select>
+        </div>
+
         <button type="submit" class="btn btn-primary">Search</button>
     }
 </div>
